requirements.txt
discord.py[voice]>=2.3.2
yt-dlp>=2023.12.30
PyNaCl>=1.5.0
python-dotenv>=1.0.0
spotipy>=2.23.0
aiohttp>=3.9.1
asyncio>=3.4.3
wavelink>=2.6.0
